Moved `wav_to_base64` function to `common.py`
